turned isolated commands into functions being called, for modularity
